consumer/requirements.txt
kafka-python
nltk
SQLAlchemy
psycopg2-binary # For PostgreSQL connectivity
textblob # <--- NEW
transformers # <--- NEW
torch # <--- NEW, for Hugging Face Transformers
